requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
google-generativeai==0.8.5
gunicorn==20.1.0
